- Merged 27155:27318 from trunk
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2007/platformopt@27319 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/csuctransform.h
@@ -950,9 +950,9 @@ public:
   inline static int UTF16Skip (const utf16_char* str, size_t maxSkip)
   {
     if (CS_UC_IS_HIGH_SURROGATE (*str))
-      return (int)(MIN(maxSkip, 2));
+      return (int)(MIN(maxSkip, (size_t)2));
     else
-      return (int)(MIN(maxSkip, 1));
+      return (int)(MIN(maxSkip, (size_t)1));
   }
   
   /**
@@ -984,7 +984,7 @@ public:
   inline static int UTF32Skip (const utf32_char* str, size_t maxSkip)
   {
     (void)str; // silence gcc
-    return (int)(MIN(maxSkip, 1));
+    return (int)(MIN(maxSkip, (size_t)1));
   }
 
   /**
